refactor: member controller 에 비즈니스 로직 제거
