+++ H-Mingle/src/main/java/com/hyundai/hmingle/service/MemberService.java
@@ -2,11 +2,13 @@ package com.hyundai.hmingle.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.hyundai.hmingle.controller.dto.request.ImageCreateRequest;
 import com.hyundai.hmingle.controller.dto.request.MemberUpdateRequest;
 import com.hyundai.hmingle.controller.dto.response.MemberGetResponse;
 import com.hyundai.hmingle.domain.member.Member;
+import com.hyundai.hmingle.support.ImageUtils;
 import com.hyundai.hmingle.mapper.dto.request.ImageUpdateMapperRequest;
 import com.hyundai.hmingle.mapper.dto.response.MemberUpdateMapperResponse;
 import com.hyundai.hmingle.repository.MemberRepository;
@@ -22,6 +24,7 @@ public class MemberService {
 
 	private final MemberRepository memberRepository;
 	private final TokenRepository tokenRepository;
+	private final ImageUtils imageUtils;
 
 	@Transactional(readOnly = true)
 	public MemberGetResponse findById(Long id, Long memberId) {
@@ -38,8 +41,9 @@ public class MemberService {
 		memberRepository.delete(savedMember.getId());
 	}
 
-	public void updateFile(Long memberId, ImageCreateRequest img) {
-		ImageUpdateMapperRequest dto = new ImageUpdateMapperRequest(memberId, img.getSaveName());
+	public void updateFile(Long memberId, MultipartFile image) {
+		ImageCreateRequest uploadImg = imageUtils.updateFile(image);
+		ImageUpdateMapperRequest dto = new ImageUpdateMapperRequest(memberId, uploadImg.getSaveName());
 
 		memberRepository.updateImg(dto);
 	}
